checking functions and ui components

+++ model/src/main/java/com/android/model/database/jobs/entity/JobRoomEntity.kt
@@ -11,7 +11,7 @@ data class JobRoomEntity(
 ) {
 
    fun mapTotJobEntity() = JobEntity(
-      id = id,
+      id = id.toInt(),
       name = name
    )
 }
